refactor(organon): extract energeia::observation submodule (#2642)
Move parateresis and mathesis tool definitions and executors to
observation.rs.
